CDPD-19596: KafkaHook implementation
This feature adds the Kafka hook capabilities. It updates the Topic
typedef and adds producer, consumer and consumergroup entity types
with relationships between them.

This new hook is able to add and delete topics when a request is captured
in Kafka, add producers, consumers and consumer groups when a client
produces or consumes.